requirements.txt
pip~=21.0.1
wheel~=0.36.2
PyYAML~=5.4.1
requests~=2.25.1
yarg~=0.1.9
certifi~=2020.12.5
chardet~=4.0.0
docopt~=0.6.2
pipreqs~=0.4.10
urllib3~=1.26.3
idna~=2.10
setuptools~=53.0.0
pynput==1.7.3
PyAutoGUI==0.9.52
PyYAML==5.4.1